pyproject.toml
[project]
name = "itau"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.124.2",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.1",
    "langchain[anthropic]>=1.1.3",
    "loguru>=0.7.3",
    "openai>=2.9.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
